refactor: Move test files to __tests__ from fixture dir
