+++ packages/artifact-testing/__tests__/relations/manyToMany.sample.ts
@@ -1,0 +1,4 @@
+import { defineCategoryFactory, definePostFactory } from "../../fixtures/relations-many-to-many/__generated__/fabbrica";
+
+export const PostFactory = definePostFactory();
+export const CategoryFactory = defineCategoryFactory();
