+++ packages/artifact-testing/__tests__/relations/oneToMany.sample.ts
@@ -1,13 +1,20 @@
-import { definePostFactory, defineUserFactory, defineReviewFactory } from "./__generated__/fabbrica";
+import {
+  definePostFactory,
+  defineUserFactory,
+  defineReviewFactory,
+} from "../../fixtures/relations-one-to-many/__generated__/fabbrica";
 
 export const UserFactory = defineUserFactory();
+
 export const PostFactory = definePostFactory();
+
 export const ReviewFactory = defineReviewFactory({
   defaultData: {
     post: PostFactory,
     reviewer: UserFactory,
   },
 });
+
 export const PostFactoryAlt = definePostFactory({
   defaultData: {
     author: UserFactory,
