+++ packages/artifact-testing/__tests__/relations/oneToOne.sample.ts
@@ -1,4 +1,4 @@
-import { defineUserFactory, defineProfileFactory } from "./__generated__/fabbrica";
+import { defineUserFactory, defineProfileFactory } from "../../fixtures/relations-one-to-one/__generated__/fabbrica";
 
 export const UserFactory = defineUserFactory();
 export const ProfileFactory = defineProfileFactory({ defaultData: { user: UserFactory } });
