--- packages/artifact-testing/fixtures/relations-many-to-many/fieldOverride.ts
@@ -1,12 +1,0 @@
-import { definePostFactory } from "./__generated__/fabbrica";
-
-export const PostFactory = definePostFactory({
-  defaultData: async () => ({
-    title: generateRandomTitle(),
-  }),
-});
-
-function generateRandomTitle() {
-  // dummy
-  return "";
-}
