--- packages/artifact-testing/fixtures/relations-many-to-many/sample.ts
@@ -1,4 +1,0 @@
-import { defineCategoryFactory, definePostFactory } from "./__generated__/fabbrica";
-
-export const PostFactory = definePostFactory();
-export const CategoryFactory = defineCategoryFactory();
