+++ packages/artifact-testing/jest.config.mjs
@@ -1,6 +1,7 @@
 export default {
   preset: "ts-jest",
-  testPathIgnorePatterns: ["<rootDir>/lib/", "<rootDir>/node_modules/"],
+  testMatch: ["**/__tests__/**/*.test.ts"],
+  testPathIgnorePatterns: ["<rootDir>/node_modules/", "<rootDir>/.built/", "<rootDir>/fixtures/"],
   transform: {
     "^.+\\.tsx?$": ["ts-jest", { diagnostics: false }],
   },
